+++ embedded-graphics/src/pixelcolor/binary_color.rs
@@ -117,10 +117,6 @@ impl BinaryColor {
 
 impl PixelColor for BinaryColor {
     type Raw = RawU1;
-
-    fn into_raw(&self) -> <Self::Raw as RawData>::Storage {
-        self.map_color(0, 1)
-    }
 }
 
 impl From<RawU1> for BinaryColor {
@@ -152,7 +148,7 @@ impl From<bool> for BinaryColor {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::pixelcolor::{Rgb565, RgbColor};
+    use crate::pixelcolor::{PixelColorExt, Rgb565, RgbColor};
 
     #[test]
     fn default_color_is_off() {