+++ embedded-graphics/src/prelude.rs
@@ -5,7 +5,7 @@ pub use crate::{
     fonts::Font,
     geometry::{Dimensions, Point, Size},
     image::{ImageDimensions, IntoPixelIter},
-    pixelcolor::{raw::RawData, GrayColor, PixelColor, RgbColor},
+    pixelcolor::{raw::RawData, GrayColor, PixelColor, PixelColorExt, RgbColor},
     primitives::Primitive,
     transform::Transform,
     DrawTarget,